@@ -23,18 +23,19 @@
     "date-fns": "^2.16.1",
     "faunadb": "^4.0.0",
     "google-trends-api": "^4.9.0",
-    "vue": "^2.6.11"
+    "vue": "^3.0.4"
   },
   "devDependencies": {
     "@types/jest": "^24.0.19",
     "@typescript-eslint/eslint-plugin": "^2.33.0",
     "@typescript-eslint/parser": "^2.33.0",
     "@vercel/node": "^1.8.4",
-    "@vue/cli-plugin-babel": "~4.5.0",
-    "@vue/cli-plugin-eslint": "~4.5.0",
-    "@vue/cli-plugin-typescript": "^4.5.7",
-    "@vue/cli-plugin-unit-jest": "^4.5.7",
-    "@vue/cli-service": "~4.5.0",
+    "@vue/cli-plugin-babel": "^4.5.9",
+    "@vue/cli-plugin-eslint": "^4.5.9",
+    "@vue/cli-plugin-typescript": "^4.5.9",
+    "@vue/cli-plugin-unit-jest": "^4.5.9",
+    "@vue/cli-service": "^4.5.9",
+    "@vue/compiler-sfc": "^3.0.4",
     "@vue/eslint-config-typescript": "^5.0.2",
     "@vue/test-utils": "^1.0.3",
     "babel-eslint": "^10.1.0",
@@ -47,8 +48,7 @@
     "sass-loader": "^10.0.3",
     "typescript": "~3.9.3",
     "vue-cli-plugin-tailwind": "^2.0.1",
-    "vue-eslint-parser": "^7.1.1",
-    "vue-template-compiler": "^2.6.11"
+    "vue-eslint-parser": "^7.1.1"
   },
   "eslintConfig": {
     "root": true,
