@@ -1,88 +1,28 @@
 <template>
-  <div @mouseenter="color = activeColor" @mouseleave="color = defaultColor">
-    <svg
-      id="svg1650"
-      xmlns:dc="http://purl.org/dc/elements/1.1/"
-      xmlns:cc="http://creativecommons.org/ns#"
-      xmlns:rdf="http://www.w3.org/1999/02/22-rdf-syntax-ns#"
-      xmlns:svg="http://www.w3.org/2000/svg"
-      xmlns="http://www.w3.org/2000/svg"
-      xmlns:sodipodi="http://sodipodi.sourceforge.net/DTD/sodipodi-0.dtd"
-      xmlns:inkscape="http://www.inkscape.org/namespaces/inkscape"
-      width="24px"
-      height="24px"
-      viewBox="0 0 11.493147 11.209467"
-      version="1.1"
-      inkscape:version="1.0.1 (c497b03c, 2020-09-10)"
-      sodipodi:docname="githubLogo.svg"
-    >
-      <defs id="defs1644" />
-      <sodipodi:namedview
-        id="base"
-        pagecolor="#ffffff"
-        bordercolor="#666666"
-        borderopacity="1.0"
-        inkscape:pageopacity="0.0"
-        inkscape:pageshadow="2"
-        inkscape:zoom="1.4755298"
-        inkscape:cx="-23.703175"
-        inkscape:cy="113.81622"
-        inkscape:document-units="mm"
-        inkscape:current-layer="layer1"
-        inkscape:document-rotation="0"
-        showgrid="false"
-        inkscape:window-width="1960"
-        inkscape:window-height="1073"
-        inkscape:window-x="0"
-        inkscape:window-y="25"
-        inkscape:window-maximized="0"
+  <div style="width: 24px; height: 24px" class="root">
+    <svg role="img" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
+      <title>GitHub icon</title>
+      <path
+        class="icon"
+        d="M12 .297c-6.63 0-12 5.373-12 12 0 5.303 3.438 9.8 8.205 11.385.6.113.82-.258.82-.577 0-.285-.01-1.04-.015-2.04-3.338.724-4.042-1.61-4.042-1.61C4.422 18.07 3.633 17.7 3.633 17.7c-1.087-.744.084-.729.084-.729 1.205.084 1.838 1.236 1.838 1.236 1.07 1.835 2.809 1.305 3.495.998.108-.776.417-1.305.76-1.605-2.665-.3-5.466-1.332-5.466-5.93 0-1.31.465-2.38 1.235-3.22-.135-.303-.54-1.523.105-3.176 0 0 1.005-.322 3.3 1.23.96-.267 1.98-.399 3-.405 1.02.006 2.04.138 3 .405 2.28-1.552 3.285-1.23 3.285-1.23.645 1.653.24 2.873.12 3.176.765.84 1.23 1.91 1.23 3.22 0 4.61-2.805 5.625-5.475 5.92.42.36.81 1.096.81 2.22 0 1.606-.015 2.896-.015 3.286 0 .315.21.69.825.57C20.565 22.092 24 17.592 24 12.297c0-6.627-5.373-12-12-12"
       />
-      <metadata id="metadata1647">
-        <rdf:RDF>
-          <cc:Work rdf:about="">
-            <dc:format>image/svg+xml</dc:format>
-            <dc:type rdf:resource="http://purl.org/dc/dcmitype/StillImage" />
-            <dc:title></dc:title>
-          </cc:Work>
-        </rdf:RDF>
-      </metadata>
-      <g
-        id="layer1"
-        inkscape:label="Layer 1"
-        inkscape:groupmode="layer"
-        transform="translate(-158.2951,-87.37741)"
-      >
-        <g
-          id="g1112"
-          transform="matrix(0.35277777,0,0,-0.35277777,164.04114,87.37741)"
-        >
-          <path
-            id="path1114"
-            d="m 0,0 c -8.995,0 -16.288,-7.293 -16.288,-16.29 0,-7.197 4.667,-13.302 11.14,-15.457 0.815,-0.149 1.112,0.354 1.112,0.786 0,0.386 -0.014,1.411 -0.022,2.77 -4.531,-0.984 -5.487,2.184 -5.487,2.184 -0.741,1.881 -1.809,2.382 -1.809,2.382 -1.479,1.011 0.112,0.991 0.112,0.991 1.635,-0.116 2.495,-1.679 2.495,-1.679 1.453,-2.489 3.813,-1.77 4.741,-1.354 0.148,1.053 0.568,1.771 1.034,2.178 -3.617,0.411 -7.42,1.809 -7.42,8.051 0,1.778 0.635,3.232 1.677,4.371 -0.168,0.412 -0.727,2.068 0.159,4.311 0,0 1.368,0.438 4.48,-1.67 1.299,0.361 2.693,0.542 4.078,0.548 1.383,-0.006 2.777,-0.187 4.078,-0.548 3.11,2.108 4.475,1.67 4.475,1.67 0.889,-2.243 0.33,-3.899 0.162,-4.311 1.044,-1.139 1.675,-2.593 1.675,-4.371 0,-6.258 -3.809,-7.635 -7.438,-8.038 0.585,-0.503 1.106,-1.497 1.106,-3.017 0,-2.177 -0.02,-3.934 -0.02,-4.468 0,-0.436 0.293,-0.943 1.12,-0.784 6.468,2.159 11.131,8.26 11.131,15.455 C 16.291,-7.293 8.997,0 0,0"
-            style="
-              fill: #1b1817;
-              fill-opacity: 1;
-              fill-rule: evenodd;
-              stroke: none;
-            "
-            :style="{ fill: color }"
-          />
-        </g>
-      </g>
     </svg>
   </div>
 </template>
 
 <script lang="ts">
-import Vue from 'vue';
-export default Vue.extend({
+import { defineComponent } from 'vue';
+
+export default defineComponent({
   name: 'GithubIcon',
-  data() {
-    return {
-      color: '#9CA3AF',
-      defaultColor: '#9CA3AF', // text-gray-400
-      activeColor: '#1b1817', // github color
-    };
-  },
 });
 </script>
+
+<style lang="postcss" scoped>
+.icon {
+  @apply fill-current text-gray-400;
+}
+.root:hover .icon {
+  @apply fill-current text-gray-800;
+}
+</style>
